+++ wms-rpc/src/main/java/com/lsh/wms/rpc/service/stock/StockQuantRpcService.java
@@ -275,16 +275,16 @@ public class StockQuantRpcService implements IStockQuantRpcService {
                 queryStockMap.put("itemId",baseinfoItem.getItemId());
                 List<StockQuant> quants =  quantService.getQuants(queryStockMap);
                 double minLifeRet = 1;
-                Map<Integer,Integer> remainShelfLifeMap = new HashMap<Integer, Integer>();
-                List<Integer> remainShelfLifeList = new ArrayList<Integer>();
+                Map<Double,Integer> remainShelfLifeMap = new HashMap<Double, Integer>();
+                List<Double> remainShelfLifeList = new ArrayList<Double>();
                 if(quants!=null){
                     for(StockQuant quant:quants){
                         Long now = DateUtils.getCurrentSeconds();
-                        int remainDate = (int)((quant.getExpireDate() - now) / (24 * 60 * 60)) ;
+                        double remainDate = ((quant.getExpireDate() - now) / (24.0 * 60.0 * 60.0)) ;
                         if(!remainShelfLifeMap.containsKey(remainDate)){
                             double lifeRet = 1;
                             if(stockQuantInfo.getShelfLife().compareTo(BigDecimal.ZERO)!=0){
-                                lifeRet = ((double)remainDate) / stockQuantInfo.getShelfLife().doubleValue();
+                                lifeRet = (remainDate) / stockQuantInfo.getShelfLife().doubleValue();
                             }
 
                             if(lifeRet < minLifeRet){
@@ -373,14 +373,14 @@ public class StockQuantRpcService implements IStockQuantRpcService {
 
                     //捡货位和区级别的都是默认为1
                     Long now = DateUtils.getCurrentSeconds();
-                    int remainDate = (int) ((quant.getExpireDate() - now) / (24 * 60 * 60));
+                    double remainDate = ((quant.getExpireDate() - now) / (24.0 * 60.0 * 60.0));
 
                     if(location.getBinUsage().compareTo(BinUsageConstant.BIN_UASGE_PICK)==0 || location.getBinUsage().compareTo(0)==0){
                         locationInfo.setMinLifeRet(1.0);
                         locationInfo.setMinLife(0);
                     }else {
 
-                        locationInfo.setMinLifeRet(item.getShelfLife().compareTo(BigDecimal.ZERO) == 0 ? 1 : ((double) remainDate) / locationInfo.getShelfLife().doubleValue());
+                        locationInfo.setMinLifeRet(item.getShelfLife().compareTo(BigDecimal.ZERO) == 0 ? 1 : (remainDate) / locationInfo.getShelfLife().doubleValue());
                         locationInfo.setMinLife(remainDate);
                     }
                 }
